src/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord==2.0.0
discord.py==2.0.1
docopt==0.6.2
frozenlist==1.3.3
gitdb==4.0.9
GitPython==3.1.29
idna==3.4
multidict==6.0.2
pipreqs==0.4.11
psutil==5.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
pywatchman==1.4.1
PyYAML==6.0
rcon==2.3.6
requests==2.28.1
six==1.16.0
smmap==5.0.0
urllib3==1.26.12
yarg==0.1.9
yarl==1.8.1
